refactor: reorganizing folder struture and renaming widgets for better clarity.
