+++ lib/src/presenter/screens/settings/widgets/settings_button_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../shared/widgets/box_shadow_widget.dart';
+import '../../../shared/ui/widgets/box_shadow_widget.dart';
 
 class SettingsButtonWidget extends StatelessWidget {
   final String text;
