+++ lib/src/presenter/screens/starting/starting_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../shared/routes/app_route_manager.dart';
 import '../../shared/constants/app_assets.dart';
-import '../../shared/widgets/will_pop_scope_widget.dart';
+import '../../shared/ui/widgets/will_pop_scope_widget.dart';
 import '../../shared/constants/space_constants.dart';
 import 'widgets/go_to_rachar_button_widget.dart';
 import 'widgets/introduction_widget.dart';
