+++ lib/src/presenter/shared/ui/widgets/drawer_widget.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 
-class CustomDrawer extends StatelessWidget {
-  const CustomDrawer({super.key});
+//! TODO: verificar a utilização deste drawer.
+class DrawerWidget extends StatelessWidget {
+  const DrawerWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
