+++ lib/src/presenter/shared/ui/widgets/slider_widget.dart
