+++ lib/src/presenter/shared/ui/widgets/subtitle_text_widget.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
-import '../constants/space_constants.dart';
+import '../../constants/space_constants.dart';
 
-class SubitleTextWidget extends StatelessWidget {
+class SubtitleTextWidget extends StatelessWidget {
   final String subtitle;
-  const SubitleTextWidget({Key? key, required this.subtitle}) : super(key: key);
+  const SubtitleTextWidget({Key? key, required this.subtitle})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) => Padding(
